% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{boxMenuOutput}
\alias{boxMenuOutput}
\title{Create a dropdown box-menu output (client side)}
\usage{
boxMenuOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic dropdown box-menu.
}
\seealso{
\code{\link{renderMenu}} for the corresponding server-side function
  and examples, and \code{\link{dropdownMenu}} for the corresponding function
  for generating static menus.

Other menu outputs: \code{\link{dropdownMenuOutput}},
  \code{\link{menuItemOutput}}, \code{\link{menuOutput}},
  \code{\link{renderMenu}}, \code{\link{sidebarMenuOutput}}
}

