% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{controlMenuOutput}
\alias{controlMenuOutput}
\title{controlMenu Output}
\usage{
controlMenuOutput(outputId, tag = tags$div)
}
\arguments{
\item{outputId}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
Create a dynamic menu output for shinydashboard (client side)
}
\details{
This can be used as a placeholder for dynamically-generated
\code{\link{controlsidebarMenu}},  \code{\link{controlsidebarMenuItem}}
Note this is the same as menuOutput but uses a different class to identify it
hence uses a different javascript function to render in shinydashboard.js
}
\seealso{
Other menu outputs: \code{\link{dropdownMenuOutput}},
  \code{\link{menuItemOutput}}, \code{\link{menuOutput}},
  \code{\link{renderControlMenu}},
  \code{\link{renderMenu}}, \code{\link{sidebarMenuOutput}}
}
