% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlSidebar.R
\name{controlsidebarMenu}
\alias{controlsidebarMenu}
\title{create a UL list with sidebar menu class}
\usage{
controlsidebarMenu(..., heading = NULL)
}
\arguments{
\item{...}{list of items to go into the menu, each an <li> element
or controlsidebarMenuItem()}
}
\description{
create a UL list with sidebar menu class
}
