% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{controlsidebartoggle}
\alias{controlsidebartoggle}
\title{create control-sidebar toggle menu item
used internally when dashboardHeader(controlsidebar = TRUE)}
\usage{
controlsidebartoggle()
}
\description{
create control-sidebar toggle menu item
used internally when dashboardHeader(controlsidebar = TRUE)
}
\keyword{internal}
