% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{CreateControlBarRight}
\alias{CreateControlBarRight}
\title{Create a dashboard control Right sidebar.}
\usage{
CreateControlBarRight(..., paneldivs = NULL)
}
\arguments{
\item{...}{Items to put in the Control Bar.}

\item{paneldivs}{Divs for nav tabs.}
}
\description{
TODO: create functions that can create sub-components in control-sidebar,
like functions defined in dashboardSidebar.R
}
