% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{createListItems}
\alias{createListItems}
\title{Creates the list items to be embeddded into the Home Tab Panel Content.}
\usage{
createListItems(Header, ProgressValue, ProgressBarClass)
}
\arguments{
\item{Header}{This will contain the header text for the list element.}

\item{ProgressValue}{This value moves the Progress bar.}

\item{ProgressBarClass}{This is used to set the Bootstrap Class for displaying the Progress Bar.}
}
\description{
It will be used to render the items present in the Home tab.
}
