% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{createSettingsTabPanel}
\alias{createSettingsTabPanel}
\title{Creates the settings Panel Div for the Nav Tabs in the Dashboard Control Bar.}
\usage{
createSettingsTabPanel(..., panelHeading)
}
\arguments{
\item{...}{Form Items to Put in the Settings Tab Content Panel.}

\item{panelHeading}{panel heading name}
}
\description{
It will contain the Tab Panel Items Specific to the General Settings Tab.
}
