% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customCheckboxInput.R
\name{customCheckboxInput}
\alias{customCheckboxInput}
\title{Create a Checkbox Custom COntrol for shinyDashboard.}
\usage{
customCheckboxInput(inputId = NULL, label = NULL, value = FALSE)
}
\arguments{
\item{inputId}{It will act as a identifier to the checkbox control element}

\item{label}{This wil Place the text along with the checkbox control}

\item{value}{it will specify true or false whether i.e checked or not}
}
\description{
It will write a default chekcbox input without any surrounding Div
}
