% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{dashboardBody}
\alias{dashboardBody}
\title{The main body of a dashboard page.}
\usage{
dashboardBody(...)
}
\arguments{
\item{...}{Items to put in the dashboard body.}
}
\description{
The main body typically contains \code{\link[=box]{box()}}es. Another common use
pattern is for the main body to contain \code{\link[=tabItems]{tabItems()}}.
}
\seealso{
\code{\link[=tabItems]{tabItems()}}, \code{\link[=box]{box()}}, \code{\link[=valueBox]{valueBox()}}.
}
