% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlSidebar.R
\name{dashboardControlSidebar}
\alias{dashboardControlSidebar}
\title{create HTML for a control-sidebar per AdminLTE theme for use on the dashboard
sidebar is an aside and included in the page
currently also requires a toggle switch in the dashboardHeader(controlsidebar = TRUE)}
\usage{
dashboardControlSidebar(...)
}
\arguments{
\item{...}{list of items to include on the sidebar}
}
\description{
create HTML for a control-sidebar per AdminLTE theme for use on the dashboard
sidebar is an aside and included in the page
currently also requires a toggle switch in the dashboardHeader(controlsidebar = TRUE)
}
