% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{dashboardFooter}
\alias{dashboardFooter}
\title{The footer of a dashboard page.}
\usage{

  dashboardFooter(mainText = HTML("<strong>Copyright &copy; 2016 <a href=\\"https://github.com/rstudio/shinydashboard\\">Shiny Dashboard</a>.</strong> All rights reserved."),
  subText = HTML("<b>Version</b> 2.3.3"))
}
\arguments{
\item{...}{Items to put in the dashboard body.}
}
\description{
The footer typically contains copyright info. Another common use pattern
is for the footer to contain dynamic message commnuncation.
}
\seealso{
\code{\link{tabItems}}, \code{\link{box}}, \code{\link{valueBox}}.
}
