% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dashboardHeader}
\alias{dashboardHeader}
\title{Create a header for a dashboard page}
\usage{
dashboardHeader(...)
}
\arguments{
\item{...}{Items to put in the header. Should be \code{\link{dropdownMenu}}s..}
}
\description{
A dashboard header can be left blank, or it can include dropdown menu items
on the right side.
}
\examples{
\donttest{
library(shiny)

# A dashboard header with 3 dropdown menus
header <- dashboardHeader(

  # Dropdown menu for messages
  dropdownMenu(type = "messages", badgeStatus = "success",
    messageItem("Support Team",
      "This is the content of a message.",
      time = "5 mins"
    ),
    messageItem("Support Team",
      "This is the content of another message.",
      time = "2 hours"
    ),
    messageItem("New User",
      "Can I get some help?",
      time = "Today"
    )
  ),

  # Dropdown menu for notifications
  dropdownMenu(type = "notifications", badgeStatus = "warning",
    notificationItem(icon = icon("users"), status = "info",
      "5 new members joined today"
    ),
    notificationItem(icon = icon("warning"), status = "danger",
      "Resource usage near limit."
    ),
    notificationItem(icon = icon("shopping-cart", lib = "glyphicon"),
      status = "success", "25 sales made"
    ),
    notificationItem(icon = icon("user", lib = "glyphicon"),
      status = "danger", "You changed your username"
    )
  ),

  # Dropdown menu for tasks, with progress bar
  dropdownMenu(type = "tasks", badgeStatus = "danger",
    progressItem(value = 20, color = "aqua",
      "Refactor code"
    ),
    progressItem(value = 40, color = "green",
      "Design new layout"
    ),
    progressItem(value = 60, color = "yellow",
      "Another task"
    ),
    progressItem(value = 80, color = "red",
      "Write documentation"
    )
  )
)

shinyApp(
  ui = dashboardPage(
    header,
    dashboardSidebar(),
    dashboardBody()
  ),
  server = function(input, output) { }
)
}
}
\seealso{
\code{\link{dropdownMenu}}
}

