% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardUser.R
\name{dashboardUser}
\alias{dashboardUser}
\title{Create a dashboard user profile.}
\usage{
dashboardUser(name = "User", image = paste0("shinydashboard", "-",
  as.character(utils::packageVersion("shinydashboard")), "/img/guang.jpg"),
  description = "Alchemist", sub_description = "Member since Aug. 2015",
  stat1 = "100 Posts", stat2 = "2048 Followers", stat3 = "1024 Following",
  btn1 = "Profile", btn2 = "Sign Out")
}
\arguments{
\item{name}{userName.}

\item{image}{user profile picture.}
}
\description{
Create a dashboard user profile.
}
\seealso{
\code{\link{userOutput}} and \code{\link{renderUser}} for dynamically-generating \code{\link{dashboardUser}}.
}
