% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{descriptionBlock}
\alias{descriptionBlock}
\title{Create a description block for the main body of a dashboard.}
\usage{
descriptionBlock(header, text, percentage, icon = NULL, width = 4,
  href = NULL, color = "green")
}
\arguments{
\item{header}{The header to display in the box. Usually a number or short text.}

\item{text}{Subtitle text.}

\item{percentage}{percentage to appear in description}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{href}{An optional URL to link to.}

\item{color}{A color for the box. Valid colors are listed in
\link{validColors}.}
}
\description{
Create a description block for the main body of a dashboard.
}
\seealso{
\code{\link{box}} for usage examples.

Other boxes: \code{\link{box}}, \code{\link{infoBox}},
  \code{\link{tabBox}}, \code{\link{valueBox}}
}

