% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dropdownMenu}
\alias{dropdownMenu}
\title{Create a dropdown menu to place in a dashboard header}
\usage{
dropdownMenu(..., type = c("messages", "notifications", "tasks"),
  badgeStatus = "none")
}
\arguments{
\item{...}{Items to put in the menu. Typically, message menus should contain
  \code{\link{messageItem}}s, notification menus should contain
  \code{\link{notificationItem}}s, and task menus should contain
  \code{\link{taskItem}}s.}

\item{type}{The type of menu. Should be one of "messages", "notifications",
"tasks".}

\item{badgeStatus}{The status of the badge. This determines the badge's
color. Valid statuses are listed in \link{validStatuses}.}
}
\description{
Create a dropdown menu to place in a dashboard header
}
\seealso{
\code{\link{dashboardHeader}} for example usage.
}

