% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{footerOutput}
\alias{footerOutput}
\title{Create a dynamic footer output for shinydashboard (client side)}
\usage{
footerOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
This can be used as a placeholder for dynamically-generated \code{\link{dashboardFooter}}.
}
\seealso{
\code{\link{renderFooter}} for the corresponding server side function
  and examples.

Other footer outputs: \code{\link{renderFooter}}
}
