% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_extensions.R
\name{getIconClass}
\alias{getIconClass}
\title{Given the name of an icon}
\usage{
getIconClass(icon)
}
\arguments{
\item{give}{the name of the icon}
}
\value{
CSS classnames, like "fa fa-dashboard" or "glyphicon glyphicon-user".
}
\description{
Given the name of an icon, like "fa-dashboard" or "glyphicon-user",
}
