% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{homeTabPanel}
\alias{homeTabPanel}
\title{Creates the Home Panel Div for the Nav Tabs in the Dashboard Control bar.}
\usage{
homeTabPanel(..., panelHeading)
}
\arguments{
\item{...}{List Items to Put in the Home Tab}

\item{panelHeading}{panel heading name}
}
\description{
It will contain the Tab Panel Items Specific to the Home Tab.
}
