% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{listItems}
\alias{listItems}
\title{Creates the list items to be embeddded into the Home Tab Panel Content.}
\usage{
listItems(header, progressValue, progressBarClass)
}
\arguments{
\item{header}{This will contain the header text for the list element.}

\item{progressValue}{This value moves the Progress bar.}

\item{progressBarClass}{This is used to set the Bootstrap Class for displaying the Progress Bar.}
}
\description{
It will be used to render the items present in the Home tab.
}
