% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_extensions.R
\name{measure_box_tabular_UI}
\alias{measure_box_tabular_UI}
\title{Create a tabular box UI}
\usage{
measure_box_tabular_UI(id, title = "", column_width = 4, offset = 0)
}
\arguments{
\item{offset}{is to set the offset of box.}

\item{Id}{is the unique id of the HTML elemnt.}

\item{Title}{is the name of the box.}

\item{ColumnWidth}{is the width of the box.}
}
\value{
<div class="col-sm-6">
<div class="box box-primary " style="background-color: #F8F8FF;">
 <div class="box-header">
 <strong>Enter Title</strong>
 </div>
 <div class="box-body">
 <div id="tableBox-tab_box" class="shiny-html-output"></div>
 </div>
 </div>
 </div>
}
\description{
Create a tabular box UI
}
