% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{menuItemOutput}
\alias{menuItemOutput}
\title{Create a sidebar menu item output (client side)}
\usage{
menuItemOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic sidebar menu item.
}
\seealso{
\code{\link[=renderMenu]{renderMenu()}} for the corresponding server-side function
and examples, and \code{\link[=menuItem]{menuItem()}} for the corresponding function
for generating static sidebar menus.

Other menu outputs:
\code{\link[=dropdownMenuOutput]{dropdownMenuOutput()}},
\code{\link[=menuOutput]{menuOutput()}},
\code{\link[=renderMenu]{renderMenu()}},
\code{\link[=sidebarMenuOutput]{sidebarMenuOutput()}}
}
\concept{menu outputs}
