% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{menuOutput}
\alias{menuOutput}
\title{Create a dynamic menu output for shinydashboard (client side)}
\usage{
menuOutput(outputId, tag = tags$li)
}
\arguments{
\item{outputId}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
This can be used as a placeholder for dynamically-generated
\code{\link[=dropdownMenu]{dropdownMenu()}}, \code{\link[=notificationItem]{notificationItem()}},
\code{\link[=messageItem]{messageItem()}}, \code{\link[=taskItem]{taskItem()}} \code{\link[=sidebarMenu]{sidebarMenu()}},
or \code{\link[=menuItem]{menuItem()}}. If called directly, you must make sure to supply
the correct type of tag. It is simpler to use the wrapper functions if
present; for example, \code{\link[=dropdownMenuOutput]{dropdownMenuOutput()}} and
\code{\link[=sidebarMenuOutput]{sidebarMenuOutput()}}.
}
\seealso{
\code{\link[=renderMenu]{renderMenu()}} for the corresponding server side function
and examples.

Other menu outputs:
\code{\link[=dropdownMenuOutput]{dropdownMenuOutput()}},
\code{\link[=menuItemOutput]{menuItemOutput()}},
\code{\link[=renderMenu]{renderMenu()}},
\code{\link[=sidebarMenuOutput]{sidebarMenuOutput()}}
}
\concept{menu outputs}
