% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{messageItem}
\alias{messageItem}
\title{Create a message item to place in a dropdown message menu}
\usage{
messageItem(
  from,
  message,
  icon = shiny::icon("user"),
  time = NULL,
  href = NULL
)
}
\arguments{
\item{from}{Who the message is from.}

\item{message}{Text of the message.}

\item{icon}{An icon tag, created by \code{\link[shiny:icon]{shiny::icon()}}.}

\item{time}{String representing the time the message was sent. Any string may
be used. For example, it could be a relative date/time like "5 minutes",
"today", or "12:30pm yesterday", or an absolute time, like "2014-12-01 13:45".
If NULL, no time will be displayed.}

\item{href}{An optional URL to link to.}
}
\description{
Create a message item to place in a dropdown message menu
}
\seealso{
\code{\link[=dashboardHeader]{dashboardHeader()}} for example usage.

Other menu items:
\code{\link[=notificationItem]{notificationItem()}},
\code{\link[=taskItem]{taskItem()}}
}
\concept{menu items}
