% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleOutput.R
\name{moduleOutput}
\alias{moduleOutput}
\title{Create a dynamic module output for shinydashboard (client side)}
\usage{
moduleOutput(outputId, tag = tags$li)
}
\arguments{
\item{outputId}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
This can be used as a placeholder for dynamically-generated
\code{\link{dashboardUser}}, \code{\link{dashboardFooter}}.
If called directly, you must make sure to supply the correct
type of tag. It is simpler to use the wrapper functions if
present; for example, \code{\link{userOutput}} and
\code{\link{footerOutput}}.
}
\seealso{
\code{\link{renderUI}} for the corresponding server side function
  and examples.

\code{\link{userOutput}} and \code{\link{footerOutput}}
  as wrapper function of \code{\link{moduleOutput}}.
}
