% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{notificationItem}
\alias{notificationItem}
\title{Create a notification item to place in a dropdown notification menu}
\usage{
notificationItem(
  text,
  icon = shiny::icon("warning"),
  status = "success",
  href = NULL
)
}
\arguments{
\item{text}{The notification text.}

\item{icon}{An icon tag, created by \code{\link[shiny:icon]{shiny::icon()}}.}

\item{status}{The status of the item This determines the item's background
color. Valid statuses are listed in \link{validStatuses}.}

\item{href}{An optional URL to link to.}
}
\description{
Create a notification item to place in a dropdown notification menu
}
\seealso{
\code{\link[=dashboardHeader]{dashboardHeader()}} for example usage.

Other menu items:
\code{\link[=messageItem]{messageItem()}},
\code{\link[=taskItem]{taskItem()}}
}
\concept{menu items}
