% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_extensions.R
\name{removable_kpi_metric_box}
\alias{removable_kpi_metric_box}
\title{Create a Removable Kpi Metrics Box}
\usage{
removable_kpi_metric_box(..., id = NULL, title = NULL, footer = NULL,
  status = success, info, solidHeader = FALSE, background = NULL,
  width = 4, height = NULL, removable = TRUE)
}
\arguments{
\item{id}{is the unique id of html elemnt.}

\item{title}{is the title of the Table box.}

\item{footer}{is the footer of the table box.}

\item{background}{is the backgroundcolor of box.}

\item{width}{id the width of the box.}

\item{height}{is the heigth of the box.}

\item{removable}{is the box should be removable.}

\item{Status}{is the status of the box.}

\item{Info}{is the information about the matrix box.}

\item{solidheader}{is the header of the table box.}
}
\description{
Create a Removable Kpi Metrics Box
}
