% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{sidebarFooter}
\alias{sidebarFooter}
\title{The footer for the Sidebar in the dashboard}
\usage{
sidebarFooter(..., height = NULL, color = NULL, backgroundColor = NULL)
}
\arguments{
\item{height}{the height as an integer, pixel e.g. 250px, or a css symbol}

\item{color}{the color of the text}

\item{backgroundColor}{the color of the background}
}
\description{
The footer typically contains info about the company or application
#' is for the footer to contain dynamic message commnuncation.
The footer is statically locate at the bottom of the sidebar
}
