% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{sidebarFooterOutput}
\alias{sidebarFooterOutput}
\title{Create a dynamic footer output for shinydashboard sidebar (client side)}
\usage{
sidebarFooterOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
This can be used as a placeholder for dynamically-generated \code{\link{dashboardSidebarFooter}}.
}
\seealso{
\code{\link{renderSideboardFooter}} for the corresponding server side function
  and examples.

Other footer outputs: \code{\link{footerOutput}},
  \code{\link{renderFooter}}
}
