% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{sidebarMenuOutput}
\alias{sidebarMenuOutput}
\title{Create a sidebar menu output (client side)}
\usage{
sidebarMenuOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic sidebar menu.
}
\seealso{
\code{\link[=renderMenu]{renderMenu()}} for the corresponding server-side function
and examples, and \code{\link[=sidebarMenu]{sidebarMenu()}} for the corresponding function
for generating static sidebar menus.

Other menu outputs:
\code{\link[=dropdownMenuOutput]{dropdownMenuOutput()}},
\code{\link[=menuItemOutput]{menuItemOutput()}},
\code{\link[=menuOutput]{menuOutput()}},
\code{\link[=renderMenu]{renderMenu()}}
}
\concept{menu outputs}
