% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{sidebarSearchForm}
\alias{sidebarSearchForm}
\title{Create a search form to place in a sidebar}
\usage{
sidebarSearchForm(
  textId,
  buttonId,
  label = "Search...",
  icon = shiny::icon("search")
)
}
\arguments{
\item{textId}{Shiny input ID for the text input box.}

\item{buttonId}{Shiny input ID for the search button (which functions like an
\code{\link[shiny:actionButton]{shiny::actionButton()}}).}

\item{label}{Text label to display inside the search box.}

\item{icon}{An icon tag, created by \code{\link[shiny:icon]{shiny::icon()}}.}
}
\description{
A search form consists of a text input field and a search button.
}
\seealso{
\code{\link[=dashboardSidebar]{dashboardSidebar()}} for example usage.

Other sidebar items:
\code{\link[=sidebarMenu]{sidebarMenu()}},
\code{\link[=sidebarUserPanel]{sidebarUserPanel()}}
}
\concept{sidebar items}
