% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\name{sidebarUserPanel}
\alias{sidebarUserPanel}
\title{A panel displaying user information in a sidebar}
\usage{
sidebarUserPanel(name, subtitle = NULL, image = NULL)
}
\arguments{
\item{name}{Name of the user.}

\item{subtitle}{Text or HTML to be shown below the name.}

\item{image}{A filename or URL to use for an image of the person. If it is a
local file, the image should be contained under the www/ subdirectory of
the application.}
}
\description{
A panel displaying user information in a sidebar
}
\seealso{
\code{\link[=dashboardSidebar]{dashboardSidebar()}} for example usage.

Other sidebar items:
\code{\link[=sidebarMenu]{sidebarMenu()}},
\code{\link[=sidebarSearchForm]{sidebarSearchForm()}}
}
\concept{sidebar items}
