% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabItem}
\alias{tabItem}
\title{One tab to put inside a tab items container}
\usage{
tabItem(tabName = NULL, ...)
}
\arguments{
\item{tabName}{The name of a tab. This must correspond to the \code{tabName}
of a \code{\link[=menuItem]{menuItem()}} or \code{\link[=menuSubItem]{menuSubItem()}}.}

\item{...}{Contents of the tab.}
}
\description{
One tab to put inside a tab items container
}
\seealso{
\code{\link[=menuItem]{menuItem()}}, \code{\link[=menuSubItem]{menuSubItem()}},
\code{\link[=tabItems]{tabItems()}}. See \code{\link[=sidebarMenu]{sidebarMenu()}} for a usage example.
}
