% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{updateBoxValue}
\alias{updateBoxValue}
\title{Update the value of an infoBox or valueBox}
\usage{
updateBoxValue(session, ...)
}
\arguments{
\item{session}{The session object that the infoBox
or valueBox belongs to}

\item{...}{name=value pairs, where name is the id
(unique identifier) of an infoBox or valueBox, and
value is the value the box should display}
}
\description{
This function allows you to update the value of
an infoBox or valuebox without re-rendering the
entire box. It is useful for streaming data or
very regularly updating values
}
\seealso{
\code{\link{infoBox}}, \code{\link{valueBox}}

Other boxes: \code{\link{box}}, \code{\link{infoBox}},
  \code{\link{tabBox}}, \code{\link{valueBox}}
}
\concept{boxes}
