% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardUser.R
\name{userOutput}
\alias{userOutput}
\title{Create a dynamic user output for shinydashboard (client side)}
\usage{
userOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
This can be used as a placeholder for dynamically-generated \code{\link{dashboardUser}}.
}
\seealso{
\code{\link{renderUser}} for the corresponding server side function
  and examples.

Other user outputs: \code{\link{renderUser}}
}
