% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validStatuses}
\alias{validStatuses}
\title{Valid statuses}
\description{
These status strings correspond to colors as defined in Bootstrap's CSS.
Although the colors can vary depending on the particular CSS selector, they
generally appear as follows:
}
\details{
\itemize{
\item \code{primary} Blue (sometimes dark blue)
\item \code{success} Green
\item \code{info} Blue
\item \code{warning} Orange
\item \code{danger} Red
}
}
\keyword{internal}
