% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardSidebar.R
\docType{data}
\name{validStyles}
\alias{validStyles}
\title{Valid Sidebar-Styles}
\description{
These are valid styles configurable on dashboard sidebar. Valid styles are
listed below.
}
\details{
\itemize{
  \item \code{style1}
  \item \code{style2}
  \item \code{style3}
}
}
\keyword{internal}
