% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxOutput.R
\name{valueBoxOutput}
\alias{valueBoxOutput}
\alias{infoBoxOutput}
\title{Create an info or value box output (client side)}
\usage{
valueBoxOutput(outputId, width = 4)

infoBoxOutput(outputId, width = 4)
}
\arguments{
\item{outputId}{Output variable name.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}
}
\description{
This is the UI-side function for creating a dynamic \code{\link{valueBox}} or
\code{\link{infoBox}}.
}
\seealso{
\code{\link{renderValueBox}} for the corresponding server-side
  function and examples.
}
