% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlSidebar.R
\name{controlsidebarMenuItem}
\alias{controlsidebarMenuItem}
\title{create items to go into controlsidebar}
\usage{
controlsidebarMenuItem(text, icon = shiny::icon("code"), href = NULL,
  heading = NA)
}
\arguments{
\item{text}{text inside <p> element}
}
\description{
create items to go into controlsidebar
}
