% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{createNavTabs}
\alias{createNavTabs}
\title{Create Nav tabs on the Shiny Dashboard Control Sidebar.}
\usage{
createNavTabs(tabID, icon = NULL)
}
\arguments{
\item{tabID}{Shiny Input ID for the Tab anchor link}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}
}
\description{
The tabs can contain tab ID which will be specific to display that tab Content in the div
below it
}
