% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardBody.R
\name{dashboardBody}
\alias{dashboardBody}
\title{The main body of a dashboard page.}
\usage{
dashboardBody(...)
}
\arguments{
\item{...}{Items to put in the dashboard body.}
}
\description{
The main body typically contains \code{\link{box}}es. Another common use
pattern is for the main body to contain \code{\link{tabItems}}.
}
\seealso{
\code{\link{tabItems}}, \code{\link{box}}, \code{\link{valueBox}}.
}
