% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{dashboardControlbar}
\alias{dashboardControlbar}
\title{Create a dashboard control Right sidebar.}
\usage{
dashboardControlbar(...)
}
\arguments{
\item{...}{Items to put in the Control Bar.}

\item{paneldivs}{Divs for nav tabs.}
}
\description{
TODO: create functions that can create sub-components in control-sidebar,
like functions defined in dashboardSidebar.R
}
