% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{dashboardFooter}
\alias{dashboardFooter}
\title{Create a footer for a dashboard page}
\usage{
dashboardFooter(..., version = " ")
}
\arguments{
\item{...}{Items to put in the footer.}

\item{version}{optional version number displayed on left side when not in mobile size}
}
\description{
A dashboard footer can be left blank, or it can include arbitrary HTML (small)
on the right side.
}
\examples{
## Footer Example
library(shiny)
copyrighttext <- function(holder=
   (if(""==Sys.getenv("USERNAME")) Sys.getenv("USER") else Sys.getenv("USERNAME"))
  ){
    paste("Copyright", format(Sys.Date(),"\%Y"),holder )
}

title <- "The Shiniest App"
footer <- dashboardFooter(
 tags$strong(title),
 HTML("|"),
 tags$i(copyrighttext()),
 version=paste("version ", packageVersion("shinydashboard"))
)
}
