% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardFooter.R
\name{dashboardFooter}
\alias{dashboardFooter}
\title{The footer of a dashboard page.}
\usage{
dashboardFooter(mainText = "", subText = NULL)
}
\arguments{
\item{mainText}{A character or html argument.  Defines the left hand part of the footer.}

\item{subText}{A character or html argument.  Defines the right hand part of the footer.}

\item{...}{Items to put in the dashboard main footer}
}
\description{
The footer typically contains copyright info. Another common use pattern
is for the footer to contain dynamic message commnuncation.
}
\seealso{
\code{\link{tabItems}}, \code{\link{box}}, \code{\link{valueBox}}.
}
