% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dashboardPage}
\alias{dashboardPage}
\title{Dashboard page}
\usage{
dashboardPage(header, sidebar, body, skin = c("blue", "black"))
}
\arguments{
\item{header}{A header created by \code{dashboardHeader}.}

\item{sidebar}{A sidebar created by \code{dashboardSidebar}.}

\item{body}{A body created by \code{dashboardBody}.}

\item{skin}{A color theme; either \code{"blue"} or \code{"black"}.}
}
\description{
This creates a dashboard page for use in a Shiny app.
}
\examples{
\donttest{
# Basic dashboard page template
library(shiny)
shinyApp(
  ui = dashboardPage(
    dashboardHeader(),
    dashboardSidebar(),
    dashboardBody()
  ),
  server = function(input, output) { }
)
}
}
\seealso{
\code{\link{dashboardHeader}}, \code{\link{dashboardSidebar}},
  \code{\link{dashboardBody}}.
}

