% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{dropdownMenuOutput}
\alias{dropdownMenuOutput}
\title{Create a dropdown menu output}
\usage{
dropdownMenuOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic dropdown menu.
}
\seealso{
\code{\link{renderDropdownMenu}} for the corresponding server-side
  function and examples.
}

