% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{dropdownMenuOutput}
\alias{dropdownMenuOutput}
\title{Create a dropdown menu output (client side)}
\usage{
dropdownMenuOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic dropdown menu.
}
\seealso{
\code{\link{renderMenu}} for the corresponding server-side function
  and examples, and \code{\link{dropdownMenu}} for the corresponding function
  for generating static menus.

Other menu outputs: 
\code{\link{menuItemOutput}()},
\code{\link{menuOutput}()},
\code{\link{renderMenu}()},
\code{\link{sidebarMenuOutput}()}
}
\concept{menu outputs}
