% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardControlbar.R
\name{formPanel}
\alias{formPanel}
\title{Creates the Form for the settings Panel which consists of checkboxes in the dashboard Control bar.}
\usage{
formPanel(Header, Description, icon = NULL)
}
\arguments{
\item{Header}{This will contain the header text for the form element.}

\item{Description}{This Provides the description for the setting element.}

\item{icon}{This will append icon to the Settings Element.}
}
\description{
It will contain Form to be processed by user
}
