% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{menuOutput}
\alias{menuOutput}
\title{Create a dynamic menu output for shinydashboard (client side)}
\usage{
menuOutput(outputId, tag = tags$li)
}
\arguments{
\item{outputId}{Output variable name.}

\item{tag}{A tag function, like \code{tags$li} or \code{tags$ul}.}
}
\description{
This can be used as a placeholder for dynamically-generated
\code{\link{dropdownMenu}}, \code{\link{notificationItem}},
\code{\link{messageItem}}, \code{\link{taskItem}} \code{\link{sidebarMenu}},
or \code{\link{menuItem}}. If called directly, you must make sure to supply
the correct type of tag. It is simpler to use the wrapper functions if
present; for example, \code{\link{dropdownMenuOutput}} and
\code{\link{sidebarMenuOutput}}.
}
\seealso{
\code{\link{renderMenu}} for the corresponding server side function
  and examples.

Other menu outputs: 
\code{\link{dropdownMenuOutput}()},
\code{\link{menuItemOutput}()},
\code{\link{renderMenu}()},
\code{\link{sidebarMenuOutput}()}
}
\concept{menu outputs}
