% Generated by roxygen2 (4.0.2): do not edit by hand
\name{messageItem}
\alias{messageItem}
\title{Create a message item to place in a dropdown message menu}
\usage{
messageItem(from, message, time, image = "foo.png")
}
\arguments{
\item{from}{Who the message is from.}

\item{message}{Text of the message.}

\item{time}{String representing the time the message was sent. Any string may
be used. For example, it could be a relative date/time like "5 minutes",
"today", or "12:30pm yesterday", or an absolute time, like "2014-12-01 13:45".}

\item{image}{the path to an image to use for the message}
}
\description{
Create a message item to place in a dropdown message menu
}
\seealso{
\code{\link{dashboardHeader}} for example usage.

Other menu.items: \code{\link{notificationItem}};
  \code{\link{taskItem}}
}

