% Generated by roxygen2 (4.0.2): do not edit by hand
\name{notificationItem}
\alias{notificationItem}
\title{Create a notification item to place in a dropdown notification menu}
\usage{
notificationItem(text, icon = shiny::icon("warning"), status = "success")
}
\arguments{
\item{text}{The notification text.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{status}{The status of the item This determines the item's background
  color. Valid statuses are listed in \code{shinydashboard:::validStatuses}.}
}
\description{
Create a notification item to place in a dropdown notification menu
}
\seealso{
\code{\link{dashboardHeader}} for example usage.

Other menu.items: \code{\link{messageItem}};
  \code{\link{taskItem}}
}

