% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{renderControlMenu}
\alias{renderControlMenu}
\title{Create dynamic menu output (server side).  This is simply an alias for
\code{\link{ shiny::renderUI}} which puts replaces contents of control menu}
\usage{
renderControlMenu(expr, env = parent.frame(), quoted = FALSE,
  outputArgs = list())
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link{HTML}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{outputArgs}{A list of arguments to be passed through to the implicit
call to \code{\link{uiOutput}} when \code{renderUI} is used in an
interactive R Markdown document.}
}
\description{
Create dynamic menu output (server side).  This is simply an alias for
\code{\link{ shiny::renderUI}} which puts replaces contents of control menu
}
\seealso{
\code{\link{controlMenuOutput}} for the corresponding client side function
  and examples.

Other menu outputs: \code{\link{controlMenuOutput}},
  \code{\link{dropdownMenuOutput}},
  \code{\link{menuItemOutput}}, \code{\link{menuOutput}},
  \code{\link{renderMenu}}, \code{\link{sidebarMenuOutput}}
}
