% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{renderDropdownMenu}
\alias{renderDropdownMenu}
\title{Create a dropdown menu output (server side; deprecated)}
\usage{
renderDropdownMenu(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link{HTML}},
or a list of such objects.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
This is the server-side function for creating a dynamic dropdown menu.
}
