% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{renderDropdownMenu}
\alias{renderDropdownMenu}
\title{Create a dropdown menu output (server side; deprecated)}
\usage{
renderDropdownMenu(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a Shiny tag object, \code{\link[shiny:reexports]{HTML()}},
or a list of such objects.}

\item{env}{The parent environment for the reactive expression. By default,
this is the calling environment, the same as when defining an ordinary
non-reactive expression. If \code{expr} is a quosure and \code{quoted} is \code{TRUE},
then \code{env} is ignored.}

\item{quoted}{If it is \code{TRUE}, then the \code{\link[=quote]{quote()}}ed value of \code{expr}
will be used when \code{expr} is evaluated. If \code{expr} is a quosure and you
would like to use its expression as a value for \code{expr}, then you must set
\code{quoted} to \code{TRUE}.}
}
\description{
This is the server-side function for creating a dynamic dropdown menu.
}
