% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinydashboard-package.r
\docType{package}
\name{shinydashboard-package}
\alias{shinydashboard}
\alias{shinydashboard-package}
\title{shinydashboard: Create Dashboards with 'Shiny'}
\description{
Create dashboards with 'Shiny'. This package provides a theme on top of 'Shiny', making it easy to create attractive dashboards.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://rstudio.github.io/shinydashboard/}
  \item Report bugs at \url{https://github.com/rstudio/shinydashboard}
}

}
\author{
\strong{Maintainer}: Winston Chang \email{winston@rstudio.com}

Authors:
\itemize{
  \item Barbara Borges Ribeiro \email{barbara@rstudio.com}
}

Other contributors:
\itemize{
  \item  RStudio [copyright holder]
  \item  Almasaeed Studio (AdminLTE theme for Bootstrap) [contributor, copyright holder]
  \item  Adobe Systems Incorporated (Source Sans Pro font) [contributor, copyright holder]
}

}
\keyword{internal}
