% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sidebarMenu}
\alias{menuItem}
\alias{menuSubItem}
\alias{sidebarMenu}
\title{Create a dashboard sidebar menu and menu items.}
\usage{
sidebarMenu(...)

menuItem(text, ..., icon = NULL, badgeLabel = NULL, badgeColor = "green",
  href = NULL, tabName = NULL)

menuSubItem(text, href = NULL, tabName = NULL)
}
\arguments{
\item{...}{For menu items, this may consist of \code{\link{menuSubItem}}s.}

\item{text}{Text to show for the menu item.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{badgeLabel}{A label for an optional badge. Usually a number or a short
word like "new".}

\item{badgeColor}{A color for the badge. Valid colors are listed in
\code{shinydashboard:::validColors}.}

\item{href}{An link address. Not compatible with \code{tabName}.}

\item{tabName}{The name of a tab that this menu item will activate. Not
compatible with \code{href}.}
}
\description{
A \code{dashboardSidebar} can contain a \code{sidebarMenu}. A
\code{sidebarMenu} contains \code{menuItem}s, and they can in turn contain
\code{menuSubItem}s.
}
\details{
Menu items (and similarly, sub-items) should have a value for either
\code{href} or \code{tabName}; otherwise the item would do nothing. If it has
a value for \code{href}, then the item will simply be a link to that value.

If a \code{menuItem} has a non-NULL \code{tabName}, then the \code{menuItem}
will behave like a tab -- in other words, clicking on the \code{menuItem}
will bring a corresponding \code{tabItem} to the front, similar to a
\code{\link[shiny]{tabPanel}}. One important difference between a
\code{menuItem} and a \code{tabPanel} is that, for a \code{menuItem}, you
must also supply a corresponding \code{tabItem} with the same value for
\code{tabName}, whereas for a \code{tabPanel}, no \code{tabName} is needed.
(This is because the structure of a \code{tabPanel} is such that the tab name
can be automatically generated.) Sub-items are also able to activate
\code{tabItem}s.

Menu items (but not sub-items) also may have an optional badge. A badge is a
colored oval containing text.
}
\seealso{
\code{\link{dashboardSidebar}} for example usage.

Other sidebar.items: \code{\link{sidebarSearchForm}};
  \code{\link{sidebarUserPanel}}
}

