% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menuOutput.R
\name{sidebarMenuOutput}
\alias{sidebarMenuOutput}
\title{Create a sidebar menu output (client side)}
\usage{
sidebarMenuOutput(outputId)
}
\arguments{
\item{outputId}{Output variable name.}
}
\description{
This is the UI-side function for creating a dynamic sidebar menu.
}
\seealso{
\code{\link{renderMenu}} for the corresponding server-side function
  and examples, and \code{\link{sidebarMenu}} for the corresponding function
  for generating static sidebar menus.

Other menu outputs: 
\code{\link{dropdownMenuOutput}()},
\code{\link{menuItemOutput}()},
\code{\link{menuOutput}()},
\code{\link{renderMenu}()}
}
\concept{menu outputs}
