% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sidebarUserPanel}
\alias{sidebarUserPanel}
\title{A panel displaying user information in a sidebar}
\usage{
sidebarUserPanel(name)
}
\arguments{
\item{name}{Name of the user.}
}
\description{
A panel displaying user information in a sidebar
}
\seealso{
\code{\link{dashboardSidebar}} for example usage.

Other sidebar.items: \code{\link{menuItem}},
  \code{\link{menuSubItem}}, \code{\link{sidebarMenu}};
  \code{\link{sidebarSearchForm}}
}

