% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tagAssert}
\alias{tagAssert}
\title{Assert that a tag has specified properties}
\usage{
tagAssert(tag, type = NULL, class = NULL)
}
\arguments{
\item{tag}{A tag object.}

\item{type}{The type of a tag, like "div", "a", "span".}

\item{class}{An HTML class.}
}
\description{
Assert that a tag has specified properties
}
\keyword{internal}

