% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tagAssert}
\alias{tagAssert}
\title{Assert that a tag has specified properties}
\usage{
tagAssert(tag, type = NULL, class = NULL, allowUI = TRUE)
}
\arguments{
\item{tag}{A tag object.}

\item{type}{The type of a tag, like "div", "a", "span".}

\item{class}{An HTML class.}

\item{allowUI}{If TRUE (the default), allow dynamic outputs generated by
\code{\link[shiny]{uiOutput}} or \code{\link[shiny]{htmlOutput}}. When a
dynamic output is provided, \code{tagAssert} won't try to validate the the
contents.}
}
\description{
Assert that a tag has specified properties
}
\keyword{internal}
