% Generated by roxygen2 (4.0.2): do not edit by hand
\name{taskItem}
\alias{taskItem}
\title{Create a task item to place in a dropdown task menu}
\usage{
taskItem(text, value = 0, color = "aqua")
}
\arguments{
\item{text}{The task text.}

\item{value}{A percent value to use for the bar.}

\item{color}{A color for the bar. Valid colors are listed in
  \link{validColors}.}
}
\description{
Create a task item to place in a dropdown task menu
}
\seealso{
\code{\link{dashboardHeader}} for example usage.

Other menu.items: \code{\link{messageItem}};
  \code{\link{notificationItem}}
}

