% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{taskItem}
\alias{taskItem}
\title{Create a task item to place in a dropdown task menu}
\usage{
taskItem(text, value = 0, color = "aqua", href = NULL)
}
\arguments{
\item{text}{The task text.}

\item{value}{A percent value to use for the bar.}

\item{color}{A color for the bar. Valid colors are listed in
\link{validColors}.}

\item{href}{An optional URL to link to.}
}
\description{
Create a task item to place in a dropdown task menu
}
\seealso{
\code{\link{dashboardHeader}} for example usage.

Other menu items: 
\code{\link{messageItem}()},
\code{\link{notificationItem}()}
}
\concept{menu items}
