% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{updateBoxValue}
\alias{updateBoxValue}
\title{Update the value of an infoBox or valueBox}
\usage{
updateBoxValue(session, inputId, value)
}
\arguments{
\item{session}{The session object that the infoBox
or valueBox belongs to}

\item{inputId}{An optional id to uniquely identify the box}

\item{value}{The value to display in the box. Usually a number or short text.}
}
\description{
This function allows you to update the value of
an infoBox or valuebox without re-rendering the
entire box. It is useful for streaming data or
very regularly updating values
}
\seealso{
\code{\link{infoBox}}, \code{\link{valueBox}}

Other boxes: \code{\link{box}}, \code{\link{infoBox}},
  \code{\link{tabBox}}, \code{\link{valueBox}}
}
\concept{boxes}
