% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{validColors}
\alias{validColors}
\title{Valid colors}
\description{
These are valid colors for various dashboard components. Valid colors are
listed below.
}
\details{
\itemize{
  \item \code{red}
  \item \code{yellow}
  \item \code{aqua}
  \item \code{blue}
  \item \code{light-blue}
  \item \code{green}
  \item \code{navy}
  \item \code{teal}
  \item \code{olive}
  \item \code{lime}
  \item \code{orange}
  \item \code{fuchsia}
  \item \code{purple}
  \item \code{maroon}
  \item \code{black}
}
}
\keyword{internal}
